separated out select-topic logic from select client
