+++ backend/src/bork-protocol/utils/topic-weights/topics.ts
@@ -240,104 +240,3 @@ export async function getAggregatedTopicWeights(
     return [];
   }
 }
-
-/**
- * Initializes topic weights if they don't exist and retrieves them
- * @param defaultTopics Default topics to use when initializing
- * @param logPrefix Prefix for log messages
- * @returns Array of topic weights or empty array if initialization fails
- */
-export async function initializeAndGetTopicWeights(
-  defaultTopics: string[],
-  logPrefix = '[Topic Processing]',
-): Promise<TopicWeightRow[]> {
-  try {
-    // First try to get aggregated weights
-    let topicWeights = await getAggregatedTopicWeights();
-
-    // Initialize if none exist
-    if (!topicWeights.length) {
-      elizaLogger.info(
-        `${logPrefix} No topic weights found, creating initial entries`,
-      );
-
-      // Create initial weight entries for each default topic
-      for (const topic of defaultTopics) {
-        const initialWeight: TopicWeightRow = {
-          id: uuidv4(),
-          topic,
-          weight: 0.5, // Start with neutral weight
-          impact_score: 0.5,
-          created_at: new Date(),
-          engagement_metrics: {
-            likes: 0,
-            retweets: 0,
-            replies: 0,
-            virality: 0.5,
-            conversionPotential: 0.5,
-            communityBuilding: 0.5,
-            thoughtLeadership: 0.5,
-          },
-          sentiment: 'neutral',
-          confidence: 0.5,
-          tweet_id: 'initial', // Mark as initial entry
-        };
-
-        try {
-          await tweetQueries.createTopicWeight(initialWeight);
-        } catch (error) {
-          elizaLogger.error(
-            `${logPrefix} Error creating initial weight for topic ${topic}:`,
-            error instanceof Error ? error.message : String(error),
-          );
-        }
-      }
-
-      elizaLogger.info(
-        `${logPrefix} Created initial entries for ${defaultTopics.length} default topics`,
-      );
-
-      // Get the aggregated weights again
-      topicWeights = await getAggregatedTopicWeights();
-
-      if (!topicWeights.length) {
-        elizaLogger.error(`${logPrefix} Failed to initialize topic weights`);
-        return []; // Return empty array if initialization failed
-      }
-    }
-
-    // Ensure all default topics are present with at least minimal weight
-    const existingTopics = new Set(topicWeights.map((tw) => tw.topic));
-    for (const topic of defaultTopics) {
-      if (!existingTopics.has(topic)) {
-        topicWeights.push({
-          id: uuidv4(),
-          topic,
-          weight: 0.3, // Lower initial weight for missing topics
-          impact_score: 0.3,
-          created_at: new Date(),
-          engagement_metrics: {
-            likes: 0,
-            retweets: 0,
-            replies: 0,
-            virality: 0.3,
-            conversionPotential: 0.3,
-            communityBuilding: 0.3,
-            thoughtLeadership: 0.3,
-          },
-          sentiment: 'neutral',
-          confidence: 0.3,
-          tweet_id: 'initial',
-        });
-      }
-    }
-
-    return topicWeights;
-  } catch (dbError) {
-    elizaLogger.error(
-      `${logPrefix} Database error getting topic weights:`,
-      dbError instanceof Error ? dbError.message : String(dbError),
-    );
-    return []; // Return empty array on error
-  }
-}
